pyproject.toml
[tool.poetry]
name = "jenkinstask"
version = "0.1.0"
description = ""
authors = ["dmitrii.mashkov <mashkovd@live.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
environs = "^9.5.0"
python-jenkins = "^1.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
